Open a search result in the transaction panel, not the editor
Tapping a search result went straight to NewEditTransactionActivity in
EDIT_ITEM mode. That screen inflates menu_new_edit_item_with_attachment,
which holds attach file and save, so a transaction found by search could be
edited but never deleted or duplicated. Every other way into a transaction
reaches TransactionItemFragment first, whether by the calendar, a per item
list, or the ACTION_ITEM_CLICK broadcast the transaction list sends, and
that fragment inflates menu_duplicate_edit_delete_item.

SearchActivity is now a MultiPanelActivity over a new
SearchMultiPanelFragment, following CalendarActivity and
TransactionListActivity. The list, the loader and the empty state come from
MultiPanelCursorListFragment and the detail panel from
MultiPanelCursorListItemFragment, so a tapped result calls showItemId and
showSecondaryPanel the way the transaction list and the calendar already do.
Delete, duplicate and the refusal to delete one half of a transfer are all
in that fragment already, so none of them is written here.

The query and the seven search filters move into the fragment, and the
loader reads them from fields. MultiPanelCursorListFragment starts the
loader while the primary panel is built, which is before setupPrimaryToolbar
inflates the search box, so that first load runs on the empty query and
matches every transaction, which is what the old screen did from
onToolbarReady. The base sets a toolbar title, so
setupPrimaryToolbar clears it before inflating the search box, which fills
the bar.

Tested on an emulator: a partial query matches and a wrong one reports
nothing found, a result opens the panel with delete on it, deleting closes
the panel and drops the row from the results with the query still in the
box, back returns to the results, and rotation keeps both the query and an
open panel. The refusal to delete one half of a transfer was not tested.
